Deleted duplicate code fragments and Added abstract class
